requirements.txt
# Scout - Merged Requirements
# Combines dependencies from both aiscore and fotmob scrapers

# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.0
requests>=2.31.0
# Optional: Better Cloudflare bypass (install with: pip install undetected-chromedriver)
# undetected-chromedriver>=3.5.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0

# Data processing (pyarrow kept for potential future use, but Parquet storage removed)
pyarrow>=14.0.0

# ClickHouse client
clickhouse-connect>=0.6.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# File locking for concurrent access protection
filelock>=3.13.0

# HTTP client with retry logic
urllib3>=2.0.0

# Retry logic
tenacity>=8.2.3

# Testing (development)
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality (development)
pylint>=3.0.3
mypy>=1.7.1
black>=23.12.0
isort>=5.13.2
flake8>=6.1.0

# Type stubs
types-requests>=2.31.0
types-PyYAML>=6.0.12
types-beautifulsoup4>=4.12.0
pandas-stubs>=2.1.0

